feat: perform sanity check in watch mode
