+++ lib/core/list-files-and-perform-sanity-checks.ts
@@ -1,0 +1,33 @@
+import glob from "glob";
+
+import { alerts } from "./alerts";
+import { ConfigOptions } from "./types";
+
+/**
+ * Return the files matching the given pattern and alert the user if only 0 or 1
+ * files matched.
+ *
+ * @param pattern the file pattern to generate type definitions for
+ * @param options the CLI options
+ */
+export function listFilesAndPerformSanityChecks(
+  pattern: string,
+  options: ConfigOptions
+): string[] {
+  // Find all the files that match the provided pattern.
+  const files = glob.sync(pattern, { ignore: options.ignore });
+
+  if (!files || !files.length) {
+    alerts.error("No files found.");
+  }
+
+  // This case still works as expected but it's easy to do on accident so
+  // provide a (hopefully) helpful warning.
+  if (files.length === 1) {
+    alerts.warn(
+      `Only 1 file found for ${pattern}. If using a glob pattern (eg: dir/**/*.scss) make sure to wrap in quotes (eg: "dir/**/*.scss").`
+    );
+  }
+
+  return files;
+}
